+++ filip/clients/ngsi_v2/cb.py
@@ -635,7 +635,7 @@ class ContextBrokerClient(BaseHttpClient):
     def delete_entity(
         self,
         entity_id: str,
-        entity_type: str,
+        entity_type: str= None,
         delete_devices: bool = False,
         iota_client: IoTAClient = None,
         iota_url: AnyHttpUrl = settings.IOTA_URL,
@@ -648,7 +648,8 @@ class ContextBrokerClient(BaseHttpClient):
             entity_id:
                 Id of the entity to be deleted
             entity_type:
-                several entities with the same entity id.
+                Entity type, to avoid ambiguity in case there are several
+                entities with the same entity id.
             delete_devices:
                 If True, also delete all devices that reference this
                 entity (entity_id as entity_name)
@@ -664,8 +665,10 @@ class ContextBrokerClient(BaseHttpClient):
         """
         url = urljoin(self.base_url, f"v2/entities/{entity_id}")
         headers = self.headers.copy()
-        params = {"type": entity_type}
-
+        if entity_type:
+            params = {'type': entity_type}
+        else:
+            params = None
         try:
             res = self.delete(url=url, params=params, headers=headers)
             if res.ok:
@@ -695,10 +698,13 @@ class ContextBrokerClient(BaseHttpClient):
                     headers=self.headers,
                 )
 
-            for device in iota_client_local.get_device_list(entity_names=[entity_id]):
-                if device.entity_type == entity_type:
+            for device in iota_client_local.get_device_list(
+                    entity_names=[entity_id]):
+                if entity_type:
+                    if device.entity_type == entity_type:
+                        iota_client_local.delete_device(device_id=device.device_id)
+                else:
                     iota_client_local.delete_device(device_id=device.device_id)
-
             iota_client_local.close()
 
     def delete_entities(self, entities: List[ContextEntity]) -> None:
@@ -743,9 +749,9 @@ class ContextBrokerClient(BaseHttpClient):
     def update_or_append_entity_attributes(
             self,
             entity_id: str,
-            entity_type: str,
             attrs: List[Union[NamedContextAttribute,
                               Dict[str, ContextAttribute]]],
+            entity_type: str = None,
             append_strict: bool = False,
             forcedUpdate: bool = False):
         """
@@ -784,6 +790,9 @@ class ContextBrokerClient(BaseHttpClient):
         params = {}
         if entity_type:
             params.update({'type': entity_type})
+        else:
+            entity_type = "dummy"
+
         options = []
         if append_strict:
             options.append("append")
@@ -896,9 +905,9 @@ class ContextBrokerClient(BaseHttpClient):
     def update_existing_entity_attributes(
             self,
             entity_id: str,
-            entity_type: str,
             attrs: List[Union[NamedContextAttribute,
                               Dict[str, ContextAttribute]]],
+            entity_type: str = None,
             forcedUpdate: bool = False,
             override_metadata: bool = False
     ):
@@ -926,7 +935,11 @@ class ContextBrokerClient(BaseHttpClient):
         """
         url = urljoin(self.base_url, f"v2/entities/{entity_id}/attrs")
         headers = self.headers.copy()
-        params = {"type": entity_type}
+        if entity_type:
+            params = {"type": entity_type}
+        else:
+            params = None
+            entity_type = "dummy"
 
         entity = ContextEntity(id=entity_id, type=entity_type)
         entity.add_attributes(attrs)
@@ -983,10 +996,10 @@ class ContextBrokerClient(BaseHttpClient):
     def replace_entity_attributes(
             self,
             entity_id: str,
-            entity_type: str,
             attrs: Union[List[Union[NamedContextAttribute,
                               Dict[str, ContextAttribute]]],
                          Dict],
+            entity_type: str = None,
             forcedUpdate: bool = False,
             key_values: bool = False,
     ):
@@ -1017,8 +1030,14 @@ class ContextBrokerClient(BaseHttpClient):
         headers = self.headers.copy()
         params = {}
         options = []
+        if entity_type:
+            params.update({"type": entity_type})
+        else:
+            entity_type = "dummy"
+
         if forcedUpdate:
             options.append("forcedUpdate")
+
         if key_values:
             options.append("keyValues")
             assert isinstance(attrs, dict)
@@ -1031,8 +1050,6 @@ class ContextBrokerClient(BaseHttpClient):
                 )
         if options:
             params.update({'options': ",".join(options)})
-        if entity_type:
-            params.update({"type": entity_type})
 
         try:
             res = self.put(
@@ -1890,8 +1907,8 @@ class ContextBrokerClient(BaseHttpClient):
         self,
         *,
         entity_id: str,
-        entity_type: str,
         command: Union[Command, NamedCommand, Dict],
+        entity_type: str = None,
         command_name: str = None,
     ) -> None:
         """